rename internal to api to reflect the functions inside are for the apis specifically
