+++ api/dgraphtypes/dgraphtypes.go
@@ -17,7 +17,7 @@ import (
 	"github.com/dgraph-io/dgo/v240/protos/api"
 	"github.com/dgraph-io/dgraph/v24/protos/pb"
 	"github.com/dgraph-io/dgraph/v24/types"
-	"github.com/hypermodeinc/modusdb/internal/structreflect"
+	"github.com/hypermodeinc/modusdb/api/structreflect"
 	"github.com/twpayne/go-geom"
 	"github.com/twpayne/go-geom/encoding/wkb"
 )
